Train notebook interface now returns the trained model - Moved train to new verb infrastructure - Train now returns the trained model object for use in notebook - Fixed: Tests won't fail on OSX's default  NOFILE rlimit of 256          files per processif the test process can change the value          itself. - Fixed: User working directory configs (./hyrax_config.toml) no longer          read when tests are being run.
