+++ src/hyrax/verbs/__init__.py
@@ -4,6 +4,7 @@
 from hyrax.verbs.database_connection import DatabaseConnection
 from hyrax.verbs.umap import Umap
 from hyrax.verbs.infer import Infer
+from hyrax.verbs.train import Train
 from hyrax.verbs.visualize import Visualize
 from hyrax.verbs.lookup import Lookup
 from hyrax.verbs.save_to_database import SaveToDatabase
@@ -20,6 +21,7 @@ __all__ = [
     "Umap",
     "Visualize",
     "Infer",
+    "Train",
     "SaveToDatabase",
     "Verb",
     "DatabaseConnection",
